Move components to "service-job" NX lib
