+++ libs/domains/service-job/feature/src/index.ts
@@ -2,6 +2,8 @@ export * from './lib/job-creation-flow/job-creation-flow'
 export * from './lib/job-creation-flow/step-introduction/step-introduction'
 export * from './lib/job-creation-flow/step-general/step-general'
 export * from './lib/job-creation-flow/step-dockerfile/step-dockerfile'
+export * from './lib/dockerfile-raw-modal/dockerfile-raw-modal'
+export * from './lib/dockerfile-settings/dockerfile-settings'
 export * from './lib/job-creation-flow/step-configure/step-configure'
 export * from './lib/job-creation-flow/step-resources/step-resources'
 export * from './lib/job-creation-flow/step-variables/step-variables'
