+++ libs/domains/services/feature/src/index.ts
@@ -63,7 +63,6 @@ export * from './lib/service-state-chip/service-state-chip'
 export * from './lib/service-clone-modal/service-clone-modal'
 export * from './lib/service-terminal/service-terminal'
 export * from './lib/need-redeploy-flag/need-redeploy-flag'
-export * from './lib/dockerfile-settings/dockerfile-settings'
 export * from './lib/confirmation-cancel-lifecycle-modal/confirmation-cancel-lifecycle-modal'
 export * from './lib/service-template-indicator/service-template-indicator'
 export * from './lib/service-avatar/service-avatar'
